+++ bin/inspectTable.js
@@ -266,7 +266,8 @@ export async function tableInspect(prompts) {
             service: 'HanaCli',
             servicePath: '/odata/v4/opensap.hana.CatalogService/',
             'openapi:url': '/odata/v4/opensap.hana.CatalogService/',
-            'openapi:diagram': true
+            'openapi:diagram': true,
+            to: 'openapi'
           })
           console.log(highlight(JSON.stringify(metadata, null, 2)))
           break
@@ -288,7 +289,8 @@ export async function tableInspect(prompts) {
             service: 'HanaCli',
             servicePath: '/odata/v4/opensap.hana.CatalogService/',
             'openapi:url': '/odata/v4/opensap.hana.CatalogService/',
-            'openapi:diagram': true
+            'openapi:diagram': true,
+            to: 'openapi'
           })
 
           let data = YAML.stringify(metadata)