+++ routes/excel.js
@@ -1,6 +1,6 @@
 // @ts-check
 import * as base from '../utils/base.js'
-import * as excel from 'node-xlsx'
+//import * as excel from 'node-xlsx'
 
 export function route (app) {
     app.get('/excel', async (req, res) => {
@@ -27,13 +27,17 @@ export function route (app) {
               out.push(innerItem)
             }
             // @ts-ignore
-            let excelOutput = excel.build([{
+            let excelOutput = ``
+            //base.error(`Excel Export temporarily disabled due to issue with install of required module in Business Application Studio`)
+            res.status(500).send(`Excel Export temporarily disabled due to issue with install of required module in Business Application Studio`)
+            
+            /*excel.build([{
               name: base.bundle.getText("gui.Results"),
               data: out
-            }])
+            }]) */
 
-			res.header("Content-Disposition", "attachment; filename=Excel.xlsx")
-			return res.type("application/vnd.ms-excel").status(200).send(excelOutput)
+			//res.header("Content-Disposition", "attachment; filename=Excel.xlsx")
+			//return res.type("application/vnd.ms-excel").status(200).send(excelOutput)
 
         } catch (error) {
             base.error(error)